2303: Add 0001-Squash-Implement-agreed-changes-to-remove-psa-truste.patch.
This PR:
    https://github.com/ARMmbed/meta-mbl/pull/619
was only merge by agreeing to remove the psa-trusted-linux-storage package
enablement, because the associated repo is not available publicly at this
time. Having it enabled means the build would break for those people without
access to psa_trusted_storage_linux! However, this patch re-instates
the changes needed to enable the package. In addition to applying this patch,
the following line should then be uncommented in imx7s-warp-mbl.conf:

    #MACHINE_FEATURES_append = " psa-storage"